Make Vector's factory methods to be user friendly: replace array arg with vararg; Fix typo in test class name.

+++ hercules-protocol/src/main/java/ru/kontur/vostok/hercules/protocol/Type.java
@@ -26,7 +26,7 @@ public enum Type {
         this.code = code;
     }
 
-    public static Type valueOf(byte code) {
+    public static Type valueOf(int code) {
         return CODES[code & 0xFF];
     }
